tickets/requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
Pillow==7.2.0
pony==0.7.13
psycopg2==2.8.5
pytz==2019.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
termcolor==1.1.0
urllib3==1.25.6
vk-api==11.5.1
